pyproject.toml
[project]
name = "emailify"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9.10"
dependencies = [
    "jinja2>=3.1.6",
    "mjml-python>=1.3.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
]
